add test in sqlProductRepository

+++ internal/core/domain/entity/product.go
@@ -1,11 +1,11 @@
 package domain
 
 type Product struct {
-	ID         int32   `db:"id"`
+	ID         int     `db:"id"`
 	Name       string  `db:"name"`
-	CategoryID int32   `db:"category_id"`
-	Length     float32 `db:"length"`
-	Price      float64 `db:"price"`
+	CategoryID int     `db:"category_id"`
+	Length     int     `db:"length"`
+	Price      int     `db:"price"`
 	Weight     float64 `db:"weight"`
 	Code       string  `db:"code"`
 }
